requirements.txt
python-telegram-bot==20.3
yt-dlp
moviepy
nest_asyncio
requests
beautifulsoup4